v1

+++ service-media/src/main/java/com/example/servicemedia/domain/xml/service/XmlDefinitionServiceImpl.java
@@ -1,9 +1,7 @@
 package com.example.servicemedia.domain.xml.service;
 
 import co.elastic.clients.elasticsearch._types.query_dsl.BoolQuery;
-import co.elastic.clients.elasticsearch._types.query_dsl.Query;
 import co.elastic.clients.elasticsearch._types.query_dsl.QueryBuilders;
-import com.example.servicemedia.domain.content.constants.ContentConstants;
 import com.example.servicemedia.domain.xml.elasticsearch.event.CreateXmlDefinitionEvent;
 import com.example.servicemedia.domain.xml.elasticsearch.event.DeleteXmlDefinitionEvent;
 import com.example.servicemedia.domain.xml.elasticsearch.event.UpdateXmlDefinitionEvent;
@@ -30,6 +28,7 @@ import org.springframework.batch.core.repository.JobRepository;
 import org.springframework.batch.core.repository.JobRestartException;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.elasticsearch.client.elc.NativeQuery;
 import org.springframework.data.elasticsearch.core.ElasticsearchOperations;
@@ -39,13 +38,15 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.TransactionSynchronization;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
+import org.springframework.util.StringUtils;
 
 import java.io.IOException;
 import java.util.Base64;
-import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
+
+import static com.example.servicemedia.util.CreatorComponent.fullTextSearchQuery;
 
 @Slf4j
 @Service
@@ -82,7 +83,7 @@ public class XmlDefinitionServiceImpl implements XmlDefinitionService {
                 .build();
         SearchHits<ElasticXmlDefinition> search = elasticsearchOperations.search(nativeQuery, ElasticXmlDefinition.class);
         Set<String> ids = search.getSearchHits().stream().map(hit -> hit.getContent().getId()).collect(Collectors.toSet());
-        return repository.findAllByIdIn(ids,pageable).map(XmlDefinitionServiceMapper::toDto);
+        return new PageImpl<>(repository.findAllByIdIn(ids,nativeQuery.getSort()),pageable,search.getTotalHits()).map(XmlDefinitionServiceMapper::toDto);
     }
 
     @Override
@@ -104,7 +105,7 @@ public class XmlDefinitionServiceImpl implements XmlDefinitionService {
         XmlDefinition xmlDefinition = new XmlDefinition();
         xmlDefinition.setXmlFile(file);
         xmlDefinition.setType(request.type());
-        xmlDefinition.setFileName(request.fileName() != null ? request.fileName() : UUID.randomUUID().toString());
+        xmlDefinition.setFileName(StringUtils.hasLength(request.fileName()) ? request.fileName() : UUID.randomUUID().toString());
 
         log.info("Saving xml definition : {}", xmlDefinition);
         repository.save(xmlDefinition);
@@ -136,7 +137,7 @@ public class XmlDefinitionServiceImpl implements XmlDefinitionService {
     public void delete(String id) {
         XmlDefinition xmlDefinition = repository.findById(id).orElseThrow(() -> new BaseException(MessageResource.NOT_FOUND, XmlDefinition.class.getSimpleName(), id));
         JobExecution execution = jobExplorer.getJobExecution(Long.parseLong(xmlDefinition.getJobExecutionId()));
-        if (execution != null) {
+        if (execution != null && StringUtils.hasLength(xmlDefinition.getJobExecutionId())) {
             JobInstance instance = execution.getJobInstance();
             log.warn("Deleting job instance : {}", instance);
             jobRepository.deleteJobInstance(instance);
@@ -165,14 +166,5 @@ public class XmlDefinitionServiceImpl implements XmlDefinitionService {
         final String xmlSignature = "<?xml";
         final String header = new String(xmlContent, 0, Math.min(5, xmlContent.length));
         return header.startsWith(xmlSignature);
-    }
-
-    private Query fullTextSearchQuery(String query) {
-        return QueryBuilders.match()
-                .field("fileName")
-                .query(query)
-                .fuzziness(ContentConstants.SEARCH_FUZZINESS)
-                .build()
-                ._toQuery();
     }
 }
